+++ andes/models/synchronous/__init__.py
@@ -4,3 +4,4 @@ Package for synchronous generators
 
 from andes.models.synchronous.gencls import GENCLS  # NOQA
 from andes.models.synchronous.genrou import GENROU  # NOQA
+from andes.models.synchronous.plbvfu1 import PLBVFU1  # NOQA